demo jstl

+++ src/main/java/ptit/d19cqcp02/hongmythaovy/model/entity/Product.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -29,7 +30,7 @@ public class Product {
     private Date productCreateDate;
 
     @Column(name = "price")
-    private Long productPrice;
+    private BigDecimal productPrice;
 
     @Column(name = "description")
     private String productDescription;
